+++ doc/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 1.3.0 (2015-XX-XX)
 ------------------
 
+* deprecated `$app['exception_handler']->disable()` in favor of `unset($app['exception_handler'])`
 * made Silex compatible with Symfony 2.7 (and keep compatibility with Symfony 2.3, 2.5, and 2.6)
 * removed deprecated TwigCoreExtension class (register the new HttpFragmentServiceProvider instead)
 * bumped minimum version of PHP to 5.3.9