+++ doc/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+1.2.4 (2015-04-11)
+------------------
+
+* fixed the exception message when mounting a collection that doesn't return a ControllerCollection
+* fixed Symfony dependencies (Silex 1.2 is not compatible with Symfony 2.7)
+
 1.2.3 (2015-01-20)
 ------------------